create products.js seed file, add export to product model file, tested working
